+++ index.ts
@@ -3,7 +3,11 @@
  */
 export const version = '__VERSION__';
 
-export { InteractiveBridge } from './src/InteractiveBridge';
+export {
+  InteractiveBridge,
+  Orientation,
+  StreamPhase,
+} from './src/InteractiveBridge';
 export { LiveryBridgeLog } from './src/livery-bridge-log/LiveryBridgeLog';
 export { LiveryBridge } from './src/LiveryBridge';
 export { MockPlayerBridge } from './src/MockPlayerBridge';