+++ src/InteractiveBridge.ts
@@ -1,5 +1,9 @@
 import { LiveryBridge } from './LiveryBridge';
 
+export type Orientation = 'landscape' | 'portrait';
+
+export type StreamPhase = 'LIVE' | 'POST' | 'PRE';
+
 /**
  * Can be used on Livery interactive layer pages to communicate with the surrounding Livery Player.
  */
@@ -26,12 +30,62 @@ export class InteractiveBridge extends LiveryBridge {
   }
 
   /**
+   * Register `handler` function to be called with `arg` and `listener` when `sendCustomCommand()` is called on
+   * other side with matching `name`.
+   *
+   * @deprecated Will be removed in the next major version. Use `registerPlayerCommand()` instead.
+   */
+  public registerCustomCommand(
+    name: string,
+    handler: (arg: unknown, listener: (value: unknown) => void) => unknown,
+  ) {
+    return super.registerCustomCommand(name, handler);
+  }
+
+  /**
+   * Register `handler` function to be called with `arg` and `listener` when `sendInteractiveCommand()` is called
+   * from the livery-player side with matching `name`.
+   */
+  public registerInteractiveCommand(
+    name: string,
+    handler: (arg: unknown, listener: (value: unknown) => void) => unknown,
+  ) {
+    this.registerCustomCommand(name, handler);
+  }
+
+  /**
+   * Returns promise of value returned by other side's custom command handler with matching `name` that is passed `arg`.
+   * Any `handler` `listener` calls will subsequently also be bridged to this `listener` callback.
+   *
+   * @deprecated Will be removed in the next major version. Use `sendPlayerCommand()` instead.
+   */
+  public sendCustomCommand<T>(
+    name: string,
+    arg?: unknown,
+    listener?: (value: T) => void,
+  ) {
+    return super.sendCustomCommand(name, arg, listener);
+  }
+
+  /**
+   * Returns promise of value returned by the livery-player's custom command handler with matching `name` that is passed `arg`.
+   * Any `handler` `listener` calls will subsequently also be bridged to this `listener` callback.
+   */
+  public sendPlayerCommand<T>(
+    name: string,
+    arg?: unknown,
+    listener?: (value: T) => void,
+  ) {
+    return this.sendCustomCommand(name, arg, listener);
+  }
+
+  /**
    * Returns promise of current LiveryPlayer window orientation (`'landscape' \| 'portrait'`)
    * and calls back `listener` with any subsequent orientations.
    */
   public subscribeOrientation(
-    listener: (orientation: 'landscape' | 'portrait') => void,
-  ) {
+    listener: (orientation: Orientation) => void,
+  ): Promise<Orientation> {
     function validate(value: unknown) {
       if (value !== 'landscape' && value !== 'portrait') {
         const strValue = JSON.stringify(value);
@@ -52,8 +106,8 @@ export class InteractiveBridge extends LiveryBridge {
    * and calls back `listener` with any subsequent phases.
    */
   public subscribeStreamPhase(
-    listener: (phase: 'LIVE' | 'POST' | 'PRE') => void,
-  ) {
+    listener: (phase: StreamPhase) => void,
+  ): Promise<StreamPhase> {
     function validate(value: unknown) {
       if (value !== 'LIVE' && value !== 'POST' && value !== 'PRE') {
         const strValue = JSON.stringify(value);
@@ -67,5 +121,21 @@ export class InteractiveBridge extends LiveryBridge {
     return this.sendCommand('subscribeStreamPhase', undefined, (value) =>
       listener(validate(value)),
     ).then(validate);
+  }
+
+  /**
+   * Unregister custom command by name.
+   *
+   * @deprecated Will be removed in the next major version. Use `unregisterInteractiveCommand()` instead.
+   */
+  public unregisterCustomCommand(name: string) {
+    super.unregisterCustomCommand(name);
+  }
+
+  /**
+   * Unregister custom interactive command by name.
+   */
+  public unregisterInteractiveCommand(name: string) {
+    return this.unregisterCustomCommand(name);
   }
 }